+++ eth_defi/tokenised_fund/supply_only.py
@@ -17,7 +17,7 @@ from eth_defi.token import TokenDetails, fetch_erc20_details
 from eth_defi.tokenised_fund.sygnum.historical import SygnumVaultHistoricalReader
 from eth_defi.tokenised_fund.vault import TokenisedFundVault
 from eth_defi.types import Percent
-from eth_defi.vault.base import TradingUniverse, VaultDepositManager, VaultFlowManager, VaultHistoricalReader, VaultInfo, VaultPortfolio, VaultSpec
+from eth_defi.vault.base import TradingUniverse, VaultFlowManager, VaultHistoricalReader, VaultInfo, VaultPortfolio, VaultSpec
 from eth_defi.vault.fee import BROKEN_FEE_DATA, FeeData
 from eth_defi.vault.lower_case_dict import LowercaseDict
 
@@ -230,14 +230,6 @@ class SupplyOnlyTokenisedFundVault(TokenisedFundVault):
 
     def get_flow_manager(self) -> VaultFlowManager:
         """Reject unsupported public flow accounting.
-
-        :raise NotImplementedError: Always.
-        """
-
-        raise NotImplementedError(self.restricted_flow_reason)
-
-    def get_deposit_manager(self) -> VaultDepositManager:
-        """Reject restricted public subscriptions and redemptions.
 
         :raise NotImplementedError: Always.
         """